requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
autoflake==0.6.6
autopep8==1.2.4
click==6.6                # via pip-tools
coverage==4.2             # via coveralls, pytest-cov
coveralls==1.1
django-autoslug==1.9.3
django-easy-select2==1.3.3
django-enumfields==0.8.2
django-environ==0.4.0
django-filter==0.15.3
django-parler-rest==1.4.2
django-parler==1.6.5      # via django-parler-rest
Django==1.10.2            # via django-environ
djangorestframework-gis==0.11
djangorestframework==3.4.7
docopt==0.6.2             # via coveralls
enum34==1.1.6             # via django-enumfields
factory-boy==2.7.0        # via pytest-factoryboy
fake-factory==0.7.2       # via factory-boy
first==2.0.1              # via pip-tools
flake8==3.0.4
inflection==0.3.1         # via pytest-factoryboy
isort==4.2.5
mccabe==0.5.2             # via flake8
pep8==1.7.0               # via autopep8
pip-tools==1.7.0
psycopg2==2.6.2
py==1.4.31                # via pytest
pycodestyle==2.0.0        # via flake8
pyflakes==1.2.3           # via autoflake, flake8
pytest-cov==2.4.0
pytest-django==3.0.0
pytest-factoryboy==1.3.0
pytest==3.0.3             # via pytest-cov, pytest-django, pytest-factoryboy
python-dateutil==2.5.3    # via fake-factory
requests==2.11.1          # via coveralls
six==1.10.0               # via django-enumfields, django-environ, fake-factory, pip-tools, python-dateutil
